Fix Photobash Crash
This was already fixed in the Github repository a while back, but I never got to publish it in the official Krita plugins, sorry!


(cherry picked from commit d57cada4d0811f243626f52fae30ba60f1961ded)

ccf20f0d fix photobash crash

+++ plugins/python/photobash_images/photobash_images_docker.py
@@ -344,11 +344,11 @@ class PhotobashDocker(DockWidget):
 
         # Scale Image
         if self.fitCanvasChecked:
-            image = QImage(photoPath).scaled(doc.width() * scale, doc.height() * scale, Qt.KeepAspectRatio, Qt.SmoothTransformation)
+            image = QImage(photoPath).scaled(int(doc.width() * scale), int(doc.height() * scale), Qt.KeepAspectRatio, Qt.SmoothTransformation)
         else:
             image = QImage(photoPath)
             # scale image
-            image = image.scaled(image.width() * scale, image.height() * scale, Qt.KeepAspectRatio, Qt.SmoothTransformation)
+            image = image.scaled(int(image.width() * scale), int(image.height() * scale), Qt.KeepAspectRatio, Qt.SmoothTransformation)
 
         # MimeData
         mimedata = QMimeData()
